refactor : home 기능 폴더 구조 수정
