+++ lib/features/home/data/models/page_result.dart
