+++ lib/features/home/data/models/room_condition.dart
