+++ lib/features/homecam/component/home_cam_view.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:flutter_mjpeg/flutter_mjpeg.dart';
 import 'package:team_project_front/core/network/dio_client.dart';
-import 'package:team_project_front/features/home/model/fever_record_data.dart';
-import 'package:team_project_front/features/home/model/room_condition.dart';
+import 'package:team_project_front/features/home/data/models/fever_record_data.dart';
+import 'package:team_project_front/features/home/data/models/room_condition.dart';
 import 'package:team_project_front/features/homecam/component/state_info_card.dart';
 import 'package:team_project_front/features/homecam/model/home_cam.dart';
 import 'dart:async';
